requirements.txt
merlinquantum
torch
pandas
openpyxl
scikit-learn
pyyaml
matplotlib
scipy
